Render a background for the open_window example (#175)
This PR adds code to render a basic gray background to the opened window in the `open_window` example.

This also helps making the example a bit more useful, since most users will want to render to their window.

And also it looks nicer. :slightly_smiling_face: 

This is done using the `softbuffer` crate, in the same manner of the `open_parented` introduced in #172.